refactor: move logic code and delete invalid test mock
